DL-OBS-Server/dl-django/dlobs/requirements.txt
# Run 'pip install -r requirements.txt' in terminal
pip==9.0.1
django==1.11.8
markdown==2.6.9
djangorestframework==3.7.3
django-filter==1.1.0
django-guardian==1.4.9
django-ipware==1.1.3
mysqlclient==1.3.12
netifaces==0.10.4
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
wsgiref==0.1.2
six==1.11.0
django-cors-headers